SIMSBIOHUB-30: Moved Landing Page actions into its own component
