+++ app/src/pages/landing/LandingPage.tsx
@@ -1,15 +1,9 @@
-import Box from '@material-ui/core/Box';
-import Button from '@material-ui/core/Button';
 import Container from '@material-ui/core/Container';
 import { makeStyles, Theme } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
-import clsx from 'clsx';
-import { AuthGuard, UnAuthGuard } from 'components/security/Guards';
-import { SYSTEM_ROLE } from 'constants/roles';
-import { AuthStateContext } from 'contexts/authStateContext';
 import BaseLayout from 'layouts/BaseLayout';
-import React, { useContext, useMemo } from 'react';
-import { Link } from 'react-router-dom';
+import React from 'react';
+import LandingActions from './LandingActions';
 
 const useStyles = makeStyles((theme: Theme) => ({
   baseLayoutContainer: {
@@ -43,58 +37,11 @@ const useStyles = makeStyles((theme: Theme) => ({
     fontSize: '1.325rem',
     lineHeight: '1.5',
     textShadow: '0px 0px 10px rgba(0,13,26,1)'
-  },
-  greeting: {
-    fontSize: '1.125rem',
-    lineHeight: '1.5'
-  },
-  heroButton: {
-    color: theme.palette.primary.main,
-    backgroundColor: '#fcba19',
-    fontWeight: 700,
-    padding: '12px 44px'
-  },
-  loginButton: {
-    textTransform: 'uppercase'
-  },
-  actionsContainer: {
-    maxWidth: '45ch',
-
-    '& p': {
-      margin: '1.5rem 0 1rem'
-    }
-  },
-  actions: {
-    display: 'flex',
-    flexFlow: 'row nowrap',
-    alignItems: 'center',
-    gap: '1rem',
-
-    '& > span': {
-      textTransform: 'uppercase'
-    }
-  },
-  heroLink: {
-    color: '#fcba19',
-    fontWeight: 700,
-    textDecoration: 'none'
-  },
-  username: {
-    textTransform: 'uppercase'
   }
 }));
 
 export const LandingPage = () => {
-  const { keycloakWrapper } = useContext(AuthStateContext);
   const classes = useStyles();
-
-  const loginUrl = useMemo(() => keycloakWrapper?.getLoginUrl(), [keycloakWrapper]);
-  const userIdentifier = keycloakWrapper?.getUserIdentifier() || '';
-
-  const hasAdministrativeRole = keycloakWrapper?.hasSystemRole([
-    SYSTEM_ROLE.DATA_ADMINISTRATOR,
-    SYSTEM_ROLE.SYSTEM_ADMIN
-  ]);
 
   return (
     <BaseLayout className={classes.baseLayoutContainer}>
@@ -106,72 +53,7 @@ export const LandingPage = () => {
           Upload and submit your species inventory project data to help understand how we can better protect and
           preserve biodiversity in British Columbia.
         </Typography>
-        <Box className={classes.actionsContainer}>
-          <UnAuthGuard>
-            <>
-              <Typography variant="body2">
-                To access this application, you must use a valid BC government-issued IDIR or BCeID account credential.
-              </Typography>
-              <Button
-                component="a"
-                href={loginUrl}
-                variant="contained"
-                className={clsx(classes.heroButton, classes.loginButton)}
-                size="large">
-                Log In
-              </Button>
-              {/**
-                * Temporarily hiding the Register link. See: https://apps.nrs.gov.bc.ca/int/jira/browse/SIMSBIOHUB-30
-                <Typography variant="body2">
-                  Don't have an account? &zwnj;
-                  <a className={classes.heroLink} href="/link-to-register-an-account">
-                    Register here.
-                  </a>
-                </Typography>
-              */}
-            </>
-          </UnAuthGuard>
-          <AuthGuard>
-            <Typography variant="body1" className={classes.greeting}>
-              <span>Welcome&nbsp;back</span>
-              {userIdentifier && (
-                <span>
-                  ,&nbsp;<strong className={classes.username}>{userIdentifier}</strong>
-                </span>
-              )}
-            </Typography>
-            <Box className={classes.actions}>
-              <Button
-                component={Link}
-                to="/admin/projects"
-                variant="contained"
-                className={classes.heroButton}
-                size="large"
-                children={<>View&nbsp;Projects</>}
-              />
-              <Typography component="span">Or</Typography>
-              {hasAdministrativeRole ? (
-                <Button
-                  component={Link}
-                  to="/admin/users"
-                  variant="contained"
-                  className={classes.heroButton}
-                  size="large"
-                  children={<>Manage&nbsp;Users</>}
-                />
-              ) : (
-                <Button
-                  component={Link}
-                  to="/admin/projects/create"
-                  variant="contained"
-                  className={classes.heroButton}
-                  size="large"
-                  children={<>Create&nbsp;a&nbsp;Project</>}
-                />
-              )}
-            </Box>
-          </AuthGuard>
-        </Box>
+        <LandingActions />
       </Container>
     </BaseLayout>
   );
